backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.128.0",
    "pydantic-settings>=2.12.0",
    "python-multipart>=0.0.22",
    "uvicorn[standard]>=0.40.0",
    "firebase-admin>=6.6.0",
    "google-genai>=1.51.0",
    "PyJWT>=2.8.0",
]
